Perf optim (#185)
* altsKeyCache as TreeSet

* set difference instead of item-wise check

* remove unused method

* quality: bytes representation

+++ src/main/scala/org/biodatageeks/sequila/pileup/model/ContigAggregate.scala
@@ -62,7 +62,7 @@ case class ContigAggregate(
     }
   }
 
-  def updateQuals(pos: Int, alt: Char, quality: Short, updateMax:Boolean = true): Unit = quals.updateQuals(pos, alt,quality, updateMax)
+  def updateQuals(pos: Int, alt: Char, quality: Byte, updateMax:Boolean = true): Unit = quals.updateQuals(pos, alt,quality, updateMax)
 
   def getTail:Tail ={
     val tailStartIndex = maxPosition - maxSeqLen
